When applied, this commit adds IO for Person, Role, and Contact as well as some bug fixes and Validations 	The commit adds the following changes: 		-Added PersonIO(Partially Finished) 		-Added RoleIO 		-Added ContactIO 		-Fixed bug where inserting same set of roles gets duplicated(Hibernate treates no id entities as new entities) 		-Fixed bug where non-required fields show null when displayed

+++ core/service/src/main/java/com/pdomingo/RoleService.java
@@ -17,6 +17,7 @@ public class RoleService {
 	public void persist(Role entity) {
 		roleDao.openCurrentSessionwithTransaction();
 		roleDao.persist(entity);
+		System.out.println("\n\t Role added! \n");
 		roleDao.closeCurrentSessionwithTransaction();
 	}
 
@@ -27,27 +28,66 @@ public class RoleService {
 	}
 
 	public void update(Role entity) {
+
 		roleDao.openCurrentSessionwithTransaction();
 		roleDao.update(entity);
+		System.out.println("\n\t Role updated! \n");
 		roleDao.closeCurrentSessionwithTransaction();
+	}
+
+	public Boolean checkIfExists(Long Id){
+		Boolean exists = true;
+
+		roleDao.openCurrentSession();
+		Role role = roleDao.findById(Id);
+		roleDao.closeCurrentSession();
+
+		if(role == null){
+			exists = false;
+		}
+
+		return exists;
 	}
 
 	public Role findById(Long id) {
 		roleDao.openCurrentSession();
 		Role role = roleDao.findById(id);
 		roleDao.closeCurrentSession();
+
 		return role;
 	}
 
 	public Role checkIfUnique(String roleText){
+		Role role = new Role();
+		try{
+			roleDao.openCurrentSession();
+			role = roleDao.findByRoleName(roleText);
 
-		return roleDao.findByRoleName(roleText);
+			if (role == null) {
+				role = new Role();
+				role.setRole(roleText);
+			}
+			roleDao.closeCurrentSession();
+		}catch(Exception e){
+
+		}
+
+		return role;
 	}
 
 	public void delete(Long id) {
-		roleDao.openCurrentSessionwithTransaction();
-		Role role = roleDao.findById(id);
-		roleDao.delete(role);
+
+
+		if(checkIfExists(id)){
+			roleDao.openCurrentSessionwithTransaction();
+			Role role = roleDao.findById(id);
+			roleDao.delete(role);
+			System.out.println("\n\t Role deleted \n");
+		}
+		else{
+			System.out.println("\n\t Role does not exist \n");
+		}
+
 		roleDao.closeCurrentSessionwithTransaction();
 	}
